DEV-1581 Stop including operations with batch FW
The existing batch operations have been moved out of the main batch
module so that they won't be included in the Docker image anymore.

This means all operations will come from the JAR the user has included
in their Docker invocation. The operations that were in the batch code
have been moved to a separate module, so that can be built as a JAR, or
the Maven approach is still available.
