+++ batch-operations/src/main/java/com/hartwig/batch/operations/AmberRerun.java
@@ -5,6 +5,7 @@ import static java.lang.String.format;
 import java.io.File;
 
 import com.hartwig.batch.BatchOperation;
+import com.hartwig.batch.OperationDescriptor;
 import com.hartwig.batch.input.InputBundle;
 import com.hartwig.batch.input.InputFileDescriptor;
 import com.hartwig.pipeline.ResultsDirectory;
