+++ batch-operations/src/main/java/com/hartwig/batch/operations/GridssRerun.java
@@ -9,6 +9,7 @@ import static com.hartwig.batch.operations.SageRerun.cramToBam;
 import java.io.File;
 
 import com.hartwig.batch.BatchOperation;
+import com.hartwig.batch.OperationDescriptor;
 import com.hartwig.batch.input.InputBundle;
 import com.hartwig.batch.input.InputFileDescriptor;
 import com.hartwig.pipeline.ResultsDirectory;
