+++ batch-operations/src/main/java/com/hartwig/batch/operations/GripssRerun.java
@@ -4,18 +4,18 @@ import java.io.File;
 import java.util.Collections;
 
 import com.hartwig.batch.BatchOperation;
+import com.hartwig.batch.OperationDescriptor;
 import com.hartwig.batch.input.ImmutableInputFileDescriptor;
 import com.hartwig.batch.input.InputBundle;
 import com.hartwig.batch.input.InputFileDescriptor;
 import com.hartwig.pipeline.ResultsDirectory;
-import com.hartwig.pipeline.calling.somatic.SagePostProcess;
+import com.hartwig.pipeline.calling.structural.gridss.stage.GridssHardFilter;
+import com.hartwig.pipeline.calling.structural.gridss.stage.GridssSomaticFilter;
 import com.hartwig.pipeline.execution.vm.BashStartupScript;
 import com.hartwig.pipeline.execution.vm.OutputFile;
 import com.hartwig.pipeline.execution.vm.OutputUpload;
 import com.hartwig.pipeline.execution.vm.RuntimeFiles;
 import com.hartwig.pipeline.execution.vm.VirtualMachineJobDefinition;
-import com.hartwig.pipeline.execution.vm.VmDirectories;
-import com.hartwig.pipeline.execution.vm.unix.UnzipToDirectoryCommand;
 import com.hartwig.pipeline.resource.RefGenomeVersion;
 import com.hartwig.pipeline.resource.ResourceFiles;
 import com.hartwig.pipeline.resource.ResourceFilesFactory;
@@ -23,58 +23,66 @@ import com.hartwig.pipeline.stages.SubStageInputOutput;
 import com.hartwig.pipeline.storage.GoogleStorageLocation;
 import com.hartwig.pipeline.storage.RuntimeBucket;
 
-public class SagePostProcessing implements BatchOperation {
+public class GripssRerun implements BatchOperation {
 
-    private static final RefGenomeVersion REF_GENOME_VERSION = RefGenomeVersion.HG38;
-
-    public static GoogleStorageLocation sageArchiveDirectory(final String set) {
-        return GoogleStorageLocation.of("hmf-sage-hg38", set, true);
+    public static GoogleStorageLocation gripssArchiveDirectory(final String set) {
+        return GoogleStorageLocation.of("hmf-gripss", set, true);
     }
 
-    public static GoogleStorageLocation hg38RemoteUnfilteredVcfArchivePath(final String set, final String sample) {
-        return GoogleStorageLocation.of("hg38-pon-pipeline-results",
-                set + "/sage/" + sample + ".sage.somatic.vcf.gz",
-                false);
+    public static GoogleStorageLocation gripssSomaticFilteredFile(final String set, final String sample) {
+        return GoogleStorageLocation.of("hmf-gripss", set + "/" + sample + ".gridss.somatic.filtered.vcf.gz", false);
+    }
+
+    public static GoogleStorageLocation gripssRecoveryFile(final String set, final String sample) {
+        return GoogleStorageLocation.of("hmf-gripss", set + "/" + sample + ".gridss.somatic.vcf.gz", false);
     }
 
     @Override
     public VirtualMachineJobDefinition execute(final InputBundle inputs, final RuntimeBucket runtimeBucket,
             final BashStartupScript startupScript, final RuntimeFiles executionFlags) {
 
-        final ResourceFiles resourceFiles = ResourceFilesFactory.buildResourceFiles(REF_GENOME_VERSION);
+        final ResourceFiles resourceFiles = ResourceFilesFactory.buildResourceFiles(RefGenomeVersion.HG19);
         final InputFileDescriptor setDescriptor = inputs.get("set");
         final String set = setDescriptor.inputValue();
-        final String sample = inputs.get("tumorSample").inputValue();
-        final InputFileDescriptor inputVcf = inputFile(setDescriptor, hg38RemoteUnfilteredVcfArchivePath(set, sample));
+        final String sample = inputs.get("tumor_sample").inputValue();
+        final InputFileDescriptor inputVcf = inputFile(setDescriptor, GridssBackport.remoteUnfilteredVcfArchivePath(set, sample));
         final InputFileDescriptor inputVcfIndex = inputVcf.index();
 
-        // Prepare SnpEff
-        startupScript.addCommand(new UnzipToDirectoryCommand(VmDirectories.RESOURCES, resourceFiles.snpEffDb()));
+        final GridssSomaticFilter somaticFilter = new GridssSomaticFilter(resourceFiles, inputVcf.localDestination());
+        final SubStageInputOutput postProcessing = somaticFilter.andThen(new GridssHardFilter())
+                .apply(SubStageInputOutput.of(sample, inputFile(inputVcf.localDestination()), Collections.emptyList()));
+
+//        // 0. Download latest jar file
+//        startupScript.addCommand(() -> format("gsutil -u hmf-crunch cp %s %s",
+//                "gs://hmf-gridss/resources/gripss.jar",
+//                "/opt/tools/gripss/" + Versions.GRIPSS + "/gripss.jar"));
 
         // 1. Download input files
         startupScript.addCommand(inputVcf::copyToLocalDestinationCommand);
         startupScript.addCommand(inputVcfIndex::copyToLocalDestinationCommand);
 
         //2. Run GRIPSS
-        SagePostProcess sagePostProcess = new SagePostProcess(sample, resourceFiles);
-        SubStageInputOutput sageOutput = sagePostProcess.apply(SubStageInputOutput.of(sample, inputFile(inputVcf.localDestination()), Collections.emptyList()));
-        startupScript.addCommands(sageOutput.bash());
+        startupScript.addCommands(postProcessing.bash());
 
         //3. Upload targeted output
-        final GoogleStorageLocation archiveStorageLocation = sageArchiveDirectory(set);
-        final OutputFile filteredOutputFile = sageOutput.outputFile();
+        final GoogleStorageLocation archiveStorageLocation = gripssArchiveDirectory(set);
+        final OutputFile filteredOutputFile = postProcessing.outputFile();
         final OutputFile filteredOutputFileIndex = filteredOutputFile.index(".tbi");
+        final OutputFile unfilteredOutputFile = somaticFilter.apply(SubStageInputOutput.empty(sample)).outputFile();
+        final OutputFile unfilteredOutputFileIndex = unfilteredOutputFile.index(".tbi");
         startupScript.addCommand(() -> filteredOutputFile.copyToRemoteLocation(archiveStorageLocation));
         startupScript.addCommand(() -> filteredOutputFileIndex.copyToRemoteLocation(archiveStorageLocation));
+        startupScript.addCommand(() -> unfilteredOutputFile.copyToRemoteLocation(archiveStorageLocation));
+        startupScript.addCommand(() -> unfilteredOutputFileIndex.copyToRemoteLocation(archiveStorageLocation));
 
         // 4. Upload output
-        startupScript.addCommand(new OutputUpload(GoogleStorageLocation.of(runtimeBucket.name(), "sage"), executionFlags));
-        return VirtualMachineJobDefinition.sageCalling(startupScript, ResultsDirectory.defaultDirectory());
+        startupScript.addCommand(new OutputUpload(GoogleStorageLocation.of(runtimeBucket.name(), "gripss"), executionFlags));
+        return VirtualMachineJobDefinition.structuralPostProcessCalling(startupScript, ResultsDirectory.defaultDirectory());
     }
 
     @Override
     public OperationDescriptor descriptor() {
-        return OperationDescriptor.of("SagePostProcessing", "Run SAGE Post Processing", OperationDescriptor.InputType.JSON);
+        return OperationDescriptor.of("GripssRerun", "Run GRIPSS", OperationDescriptor.InputType.JSON);
     }
 
     private OutputFile inputFile(final String file) {
