+++ batch-operations/src/main/java/com/hartwig/batch/operations/SambambaCramaBam.java
@@ -3,6 +3,7 @@ package com.hartwig.batch.operations;
 import java.io.File;
 
 import com.hartwig.batch.BatchOperation;
+import com.hartwig.batch.OperationDescriptor;
 import com.hartwig.batch.input.InputBundle;
 import com.hartwig.batch.input.InputFileDescriptor;
 import com.hartwig.pipeline.ResultsDirectory;
