+++ batch-operations/src/main/java/com/hartwig/batch/operations/rna/RnaIsofox.java
@@ -3,9 +3,9 @@ package com.hartwig.batch.operations.rna;
 import static java.lang.String.format;
 
 import com.hartwig.batch.BatchOperation;
+import com.hartwig.batch.OperationDescriptor;
 import com.hartwig.batch.input.InputBundle;
 import com.hartwig.batch.input.InputFileDescriptor;
-import com.hartwig.batch.operations.OperationDescriptor;
 import com.hartwig.pipeline.ResultsDirectory;
 import com.hartwig.pipeline.execution.vm.Bash;
 import com.hartwig.pipeline.execution.vm.BashStartupScript;
