+++ batch-operations/src/main/java/com/hartwig/batch/operations/rna/RnaRsem.java
@@ -11,9 +11,9 @@ import java.util.stream.Collectors;
 
 import com.google.common.collect.Lists;
 import com.hartwig.batch.BatchOperation;
+import com.hartwig.batch.OperationDescriptor;
 import com.hartwig.batch.input.InputBundle;
 import com.hartwig.batch.input.InputFileDescriptor;
-import com.hartwig.batch.operations.OperationDescriptor;
 import com.hartwig.pipeline.ResultsDirectory;
 import com.hartwig.pipeline.calling.command.VersionedToolCommand;
 import com.hartwig.pipeline.execution.vm.Bash;
@@ -21,15 +21,18 @@ import com.hartwig.pipeline.execution.vm.BashStartupScript;
 import com.hartwig.pipeline.execution.vm.ImmutableVirtualMachineJobDefinition;
 import com.hartwig.pipeline.execution.vm.OutputUpload;
 import com.hartwig.pipeline.execution.vm.RuntimeFiles;
-import com.hartwig.pipeline.execution.vm.SambambaCommand;
 import com.hartwig.pipeline.execution.vm.VirtualMachineJobDefinition;
 import com.hartwig.pipeline.execution.vm.VirtualMachinePerformanceProfile;
 import com.hartwig.pipeline.execution.vm.VmDirectories;
 import com.hartwig.pipeline.storage.GoogleStorageLocation;
 import com.hartwig.pipeline.storage.RuntimeBucket;
-import com.hartwig.pipeline.tools.Versions;
 
-public class RnaStarMapping implements BatchOperation {
+public class RnaRsem implements BatchOperation {
+
+    private static final String RSEM_TOOL = "rsem";
+    private static final String RSEM_EXPRESSION_CMD = "rsem-calculate-expression";
+    private static final String RSEM_GENE_INDEX_DIR = "rsem_gene_index";
+    private static final String RSEM_GENE_INDEX = "human_gencode";
 
     private static final String REF_GENCODE_37 = "hs37d5_GENCODE19";
     private static final String REF_LOCATION = "gs://isofox-resources";
@@ -65,21 +68,27 @@ public class RnaStarMapping implements BatchOperation {
             startupScript.addCommand(() -> format("gsutil -u hmf-crunch cp %s %s", fastqFile, VmDirectories.INPUT));
         }
 
+        // download the executables
+        startupScript.addCommand(() -> format("gsutil -u hmf-crunch cp -r %s/%s %s",
+                REF_LOCATION, RSEM_TOOL, VmDirectories.TOOLS));
+
+        startupScript.addCommand(() -> format("chmod a+x %s/%s/*", VmDirectories.TOOLS, RSEM_TOOL));
+
         // locate the FASTQ files for reads 1 and 2
         final String r1Files = format("$(ls %s/*_R1* | tr '\\n' ',')", VmDirectories.INPUT);
         final String r2Files = format("$(ls %s/*_R2* | tr '\\n' ',')", VmDirectories.INPUT);
 
-        // copy reference files for STAR
+        // download reference files
         startupScript.addCommand(() -> format("gsutil -u hmf-crunch cp -r %s/%s %s", REF_LOCATION, REF_GENCODE_37, VmDirectories.INPUT));
-
-        final String refGenomeDir = String.format("%s/%s", VmDirectories.INPUT, REF_GENCODE_37);
-
-        // logging
-        final String threadCount = Bash.allCpus();
+        startupScript.addCommand(() -> format("gsutil -u hmf-crunch cp -r %s/%s %s", REF_LOCATION, RSEM_GENE_INDEX_DIR, VmDirectories.INPUT));
 
         startupScript.addCommand(() -> format("cd %s", VmDirectories.OUTPUT));
 
-        // run the STAR mapper
+        // run STAR with transcriptome mapping
+        final String refGenomeDir = String.format("%s/%s", VmDirectories.INPUT, REF_GENCODE_37);
+
+        final String threadCount = Bash.allCpus();
+
         final String[] starArgs = {"--runThreadN", threadCount, "--genomeDir", refGenomeDir, "--genomeLoad", "NoSharedMemory",
                 "--readFilesIn", r1Files, r2Files, "--readFilesCommand", "zcat", "--outSAMtype", "BAM", "Unsorted",
                 "--outSAMunmapped", "Within", "--outBAMcompression", "0", "--outSAMattributes", "All",
@@ -87,65 +96,49 @@ public class RnaStarMapping implements BatchOperation {
                 "--chimSegmentMin", "10", "--chimOutType", "WithinBAM", "SoftClip", "--chimJunctionOverhangMin", "10", "--chimSegmentReadGapMax", "3",
                 "--chimScoreMin", "1", "--chimScoreDropMax", "30", "--chimScoreJunctionNonGTAG", "0", "--chimScoreSeparation", "1",
                 "--outFilterScoreMinOverLread", "0.33", "--outFilterMatchNminOverLread", "0.33", "--outFilterMatchNmin", "35",
-                "--alignSplicedMateMapLminOverLmate", "0.33", "--alignSplicedMateMapLmin", "35", "--alignSJstitchMismatchNmax", "5", "-1", "5", "5"};
+                "--alignSplicedMateMapLminOverLmate", "0.33", "--alignSplicedMateMapLmin", "35", "--alignSJstitchMismatchNmax", "5", "-1", "5", "5",
+                "--quantMode", "TranscriptomeSAM"}; // key line for RSEM;
 
         startupScript.addCommand(new VersionedToolCommand("star", "STAR", "2.7.3a", starArgs));
 
-        final String bamFile = "Aligned.out.bam";
+        // key output file is 'Aligned.toTranscriptome.out.bam'
 
-        // sort the BAM
-        final String sortedBam = sampleId + ".sorted.bam";
+        // ./tools/RSEM-1.3.3/rsem-calculate-expression --alignments --paired-end
+        // ./runs/CPCT02020378T/Aligned.toTranscriptome.out.bam
+        // ./ref/rsem_gene_index/human_gencode
+        // CPCT02020378T.rsem -p 6 &
 
-        final String[] sortArgs = {"sort", "-@", threadCount, "-m", "2G", "-T", "tmp", "-O", "bam", bamFile, "-o", sortedBam};
+        final String transcriptomeBam = "Aligned.toTranscriptome.out.bam";
 
-        startupScript.addCommand(new VersionedToolCommand("samtools", "samtools", Versions.SAMTOOLS, sortArgs));
+        // TMP: copy transcriptome BAM to the bucket
+        // startupScript.addCommand(() -> format("gsutil -m cp %s/%s gs://rna-cohort/%s/rsem/", VmDirectories.OUTPUT, transcriptomeBam, sampleId));
 
-        // mark duplicate fragment reads within the BAM
-        final String sortedDedupedBam = sampleId + ".sorted.dups.bam";
+        final String rsemGeneIndex = String.format("%s/%s/%s", VmDirectories.INPUT, RSEM_GENE_INDEX_DIR, RSEM_GENE_INDEX);
+        final String outputPrefix = String.format("%s.rsem", sampleId);
 
-        final String[] dupArgs = {"markdup", "-t", threadCount, "--overflow-list-size=45000000", sortedBam, sortedDedupedBam};
+        // run RSEM
+        StringBuilder rsemArgs = new StringBuilder();
+        rsemArgs.append(" --alignments");
+        rsemArgs.append(" --paired-end");
+        rsemArgs.append(String.format(" %s", transcriptomeBam));
+        rsemArgs.append(String.format(" %s", rsemGeneIndex));
+        rsemArgs.append(String.format(" %s", outputPrefix));
+        rsemArgs.append(String.format(" -p %s", threadCount));
 
-        startupScript.addCommand(new SambambaCommand(dupArgs));
+        // run RSEM transcript expression calcs
+        startupScript.addCommand(() -> format("%s/%s/%s %s", VmDirectories.TOOLS, RSEM_TOOL, RSEM_EXPRESSION_CMD, rsemArgs.toString()));
 
-        final String[] indexArgs = {"index", sortedDedupedBam};
+        startupScript.addCommand(() -> format("mv %s.rsem.genes.results %s.rsem.gene_data.tsv", sampleId, sampleId));
+        startupScript.addCommand(() -> format("mv %s.rsem.isoforms.results %s.rsem.trans_data.tsv", sampleId, sampleId));
 
-        startupScript.addCommand(new VersionedToolCommand("samtools", "samtools", Versions.SAMTOOLS, indexArgs));
-
-        // clean up intermediary BAMs
-        startupScript.addCommand(() -> format("rm -f %s", bamFile));
-        startupScript.addCommand(() -> format("rm -f %s", sortedBam));
-
-        final String starStats = "Log.final.out";
-        final String statsFile = sampleId + "." + starStats;
-        startupScript.addCommand(() -> format("mv %s %s", starStats, statsFile));
-
-        // run QC stats on the fast-Qs as well
-        /*
-        final String fastqcOutputDir = format("%s/fastqc", VmDirectories.OUTPUT);
-        startupScript.addCommand(() -> format("mkdir %s", fastqcOutputDir));
-
-        final String allFastQs = format("%s/*gz", VmDirectories.INPUT);
-        final String[] fastqcArgs = {"-o", fastqcOutputDir, allFastQs};
-
-        // TEMP until reimage has taken place
-        // startupScript.addCommand(() -> format("chmod a+x /opt/tools/fastqc/0.11.4/fastqc"));
-
-        startupScript.addCommand(new VersionedToolCommand("fastqc", "fastqc", "0.11.4", fastqcArgs));
-        */
-
-        // upload the results
-        startupScript.addCommand(new OutputUpload(GoogleStorageLocation.of(bucket.name(), "star"), executionFlags));
+        startupScript.addCommand(new OutputUpload(GoogleStorageLocation.of(bucket.name(), "rsem"), executionFlags));
 
         // copy results to rna-analysis location on crunch
+        startupScript.addCommand(() -> format("gsutil -m cp %s/*tsv gs://rna-cohort/%s/rsem/", VmDirectories.OUTPUT, sampleId));
 
-
-        startupScript.addCommand(() -> format("gsutil -m cp %s/* gs://rna-cohort/%s/", VmDirectories.OUTPUT, sampleId));
-
-        // startupScript.addCommand(new OutputUpload(GoogleStorageLocation.of("rna-cohort", sampleId), executionFlags));
-
-        return ImmutableVirtualMachineJobDefinition.builder().name("rna-star-mapping").startupCommand(startupScript)
+        return ImmutableVirtualMachineJobDefinition.builder().name("rna-rsem").startupCommand(startupScript)
                 .namespacedResults(ResultsDirectory.defaultDirectory()).workingDiskSpaceGb(500)
-                .performanceProfile(VirtualMachinePerformanceProfile.custom(12, 48)).build();
+                .performanceProfile(VirtualMachinePerformanceProfile.custom(12, 36)).build();
     }
 
     private final List<String> getSampleFastqFileList(final String sampleId, final String fastqFilelist)
@@ -177,7 +170,7 @@ public class RnaStarMapping implements BatchOperation {
 
     @Override
     public OperationDescriptor descriptor() {
-        return OperationDescriptor.of("RnaStarMapping", "Generate BAMs from RNA FASTQs",
+        return OperationDescriptor.of("RnaRsem", "Run RSEM to calculate TPM from RNA FASTQs",
                 OperationDescriptor.InputType.FLAT);
     }
