+++ batch-operations/src/test/java/com/hartwig/batch/operations/PurpleRerunTest.java
