+++ batch/src/main/java/com/hartwig/batch/BatchOperation.java
@@ -1,7 +1,6 @@
 package com.hartwig.batch;
 
 import com.hartwig.batch.input.InputBundle;
-import com.hartwig.batch.operations.OperationDescriptor;
 import com.hartwig.pipeline.execution.vm.BashStartupScript;
 import com.hartwig.pipeline.execution.vm.RuntimeFiles;
 import com.hartwig.pipeline.execution.vm.VirtualMachineJobDefinition;
