+++ batch/src/main/java/com/hartwig/batch/input/InputParserProvider.java
@@ -1,7 +1,7 @@
 package com.hartwig.batch.input;
 
 import com.hartwig.batch.BatchOperation;
-import com.hartwig.batch.operations.OperationDescriptor;
+import com.hartwig.batch.OperationDescriptor;
 
 public class InputParserProvider {
     public InputParser from(BatchOperation batchOperation) {
